fix: keep Linux control lengths type-correct

+++ src/net/quic.rs
@@ -598,7 +598,7 @@ fn gro_segment_size(header: &libc::msghdr) -> Option<usize> {
     unsafe {
         let control = libc::CMSG_FIRSTHDR(header);
         let required = libc::CMSG_LEN(std::mem::size_of::<u16>() as _) as usize;
-        let control_len = header.msg_controllen as usize;
+        let control_len = header.msg_controllen;
         if control.is_null()
             || control_len < required
             || (*control).cmsg_level != libc::IPPROTO_UDP
